added API to get MetricRecorder in LoadBalancer.Helper and add MetricSink to ManagedChannelBuilder
