+++ api/src/main/java/io/grpc/MetricRecorder.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.grpc.internal;
+package io.grpc;
 
 import io.grpc.DoubleCounterMetricInstrument;
 import io.grpc.DoubleHistogramMetricInstrument;
@@ -22,85 +22,54 @@ import io.grpc.Internal;
 import io.grpc.LongCounterMetricInstrument;
 import io.grpc.LongHistogramMetricInstrument;
 import java.util.List;
-import java.util.Set;
 
 /**
- * An internal interface representing a receiver or aggregator of gRPC metrics data.
+ * An interface used for recording gRPC metrics. Implementations of this interface are responsible
+ * for collecting and potentially reporting metrics from various gRPC components.
  */
 @Internal
-public interface MetricSink {
-
+public interface MetricRecorder {
   /**
-   * Returns a set of names for the metrics that are currently enabled for this `MetricSink`.
+   * Records a value for a double-precision counter metric instrument.
    *
-   * @return A set of enabled metric names.
-   */
-  Set<String> getEnabledMetrics();
-
-  /**
-   * Returns a list of label names that are considered optional for metrics collected by this
-   * `MetricSink`.
-   *
-   * @return A list of optional label names.
-   */
-  List<String> getOptionalLabels();
-
-  /**
-   * Returns a list of metric measures used to record metric values. These measures are created
-   * based on registered metrics (via MetricInstrumentRegistry) and are ordered according to their
-   * registration sequence.
-   *
-   * @return A list of metric measures.
-   */
-  List<Object> getMetricsMeasures();
-
-  /**
-   * Records a value for a double-precision counter associated with specified metric instrument.
-   *
-   * @param metricInstrument The counter metric instrument identifies metric measure to record.
+   * @param metricInstrument The counter metric instrument to record the value against.
    * @param value The value to record.
    * @param requiredLabelValues A list of required label values for the metric.
    * @param optionalLabelValues A list of additional, optional label values for the metric.
    */
   default void recordDoubleCounter(DoubleCounterMetricInstrument metricInstrument, double value,
-      List<String> requiredLabelValues, List<String> optionalLabelValues) {
-  }
+      List<String> requiredLabelValues, List<String> optionalLabelValues) {}
 
   /**
-   * Records a value for a long valued counter metric associated with specified metric instrument.
+   * Records a value for a long valued counter metric instrument.
    *
-   * @param metricInstrument The counter metric instrument identifies metric measure to record.
+   * @param metricInstrument The counter metric instrument to record the value against.
    * @param value The value to record.
    * @param requiredLabelValues A list of required label values for the metric.
    * @param optionalLabelValues A list of additional, optional label values for the metric.
    */
   default void recordLongCounter(LongCounterMetricInstrument metricInstrument, long value,
-      List<String> requiredLabelValues, List<String> optionalLabelValues) {
-  }
+      List<String> requiredLabelValues, List<String> optionalLabelValues) {}
 
   /**
-   * Records a value for a double-precision histogram metric associated with specified metric
-   * instrument.
+   * Records a value for a double-precision histogram metric instrument.
    *
-   * @param metricInstrument The histogram metric instrument identifies metric measure to record.
+   * @param metricInstrument The histogram metric instrument to record the value against.
    * @param value The value to record.
    * @param requiredLabelValues A list of required label values for the metric.
    * @param optionalLabelValues A list of additional, optional label values for the metric.
    */
   default void recordDoubleHistogram(DoubleHistogramMetricInstrument metricInstrument, double value,
-      List<String> requiredLabelValues, List<String> optionalLabelValues) {
-  }
+      List<String> requiredLabelValues, List<String> optionalLabelValues) {}
 
   /**
-   * Records a value for a long valued histogram metric associated with specified metric
-   * instrument.
+   * Records a value for a long valued histogram metric instrument.
    *
-   * @param metricInstrument The histogram metric instrument identifies metric measure to record.
+   * @param metricInstrument The histogram metric instrument to record the value against.
    * @param value The value to record.
    * @param requiredLabelValues A list of required label values for the metric.
    * @param optionalLabelValues A list of additional, optional label values for the metric.
    */
   default void recordLongHistogram(LongHistogramMetricInstrument metricInstrument, long value,
-      List<String> requiredLabelValues, List<String> optionalLabelValues) {
-  }
+      List<String> requiredLabelValues, List<String> optionalLabelValues) {}
 }
